% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Calculates the confidence interval for the given vector x's statistic of interest (fun)}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations to take samples}

\item{x}{given input data}

\item{fun}{statistic of interest (mean, var, etc)}

\item{alpha}{used for confidence interval}

\item{cx}{character expansion factor for text}

\item{...}{extra arguments for histogram creation}
}
\value{
named list including confidence interval ci, the function fun, and the given data x
}
\description{
Calculates the confidence interval for the given vector x's statistic of interest (fun)
}
\examples{
myboot2(x = c(1,1,1,2,2,2,2,3,3,3,4,4), fun = "median")

}
