% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Calculates the (1-a)100\% confidence interval for the mean of a single sample x}
\usage{
myci(x, a = 0.05)
}
\arguments{
\item{x}{the sample set}

\item{a}{alpha in the (1-a)100\% confidence interval formula}
}
\value{
A named list containing the lower and upper bounds of the interval, as well as the given alpha value
}
\description{
Calculates the (1-a)100\% confidence interval for the mean of a single sample x
}
\examples{
myci(c(1,2,3,4), a = 0.10)

}
