% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{Repeatedly collect samples from a uniform distribution and create a histogram
of their sum distribution in order to demonstrate the
Central Limit Theorem (CLT).}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{Number of samples to collect each iteration}

\item{iter}{Number of iterations to perform and show in the histogram}

\item{a}{Lower limit of uniform distribution}

\item{b}{Upper limit of uniform distribution}
}
\value{
Nothing
}
\description{
Repeatedly collect samples from a uniform distribution and create a histogram
of their sum distribution in order to demonstrate the
Central Limit Theorem (CLT).
}
\examples{
mycltu(n=20,iter=100000)

}
