% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Uses the provided L function to find the max likelihood estimate for some parameter graphically}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{The L function to find the MLE}

\item{x}{The values used to find the estimate}

\item{param}{A sequence for the unknown parameter}

\item{...}{Additional parameters for plotting}
}
\value{
A named list that includes the estimated parameter. Also plots a graph
}
\description{
Uses the provided L function to find the max likelihood estimate for some parameter graphically
}
