% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Displays the normal curve, with the shaded area from -Inf to x=a
and calculates the area}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean of the normal curve}

\item{sigma}{Standard deviation of the normal curve}

\item{a}{Function returns P(X<=a)}
}
\value{
Area of the shaded region, the probability of X <= a. Rounded to 4 digits
}
\description{
Displays the normal curve, with the shaded area from -Inf to x=a
and calculates the area
}
\examples{
myncurve(mu = 10, sigma = 5, a = 6)

}
