% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mySample.R
\name{mysample}
\alias{mysample}
\title{Takes and graphs a sample of numbers from one to ten, repeating multiple times.}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{number of samples to take}

\item{iter}{number of iterations}

\item{time}{time for R to suspend after generating each graph to reduce strain on the system}
}
\value{
NULL This function does not return anything
}
\description{
Takes and graphs a sample of numbers from one to ten, repeating multiple times.
}
\examples{
  mysample(n=1000, iter=30)

}
