% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Solves the "airline overbooking problem"}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of available seats (integer)}

\item{gamma}{Acceptable overbooking rate (between 0.0 and 1.0)}

\item{p}{Expected rate of attendance (between 0.0 and 1.0)}
}
\value{
Named list containing number of tickets calculated with continuous distribution (cn), discrete (dn), and the given inputs
}
\description{
Given the number of seats available, acceptable overbooking rate, and the
expected attendance rate, calculates the number of tickets that should be sold.
Calculates the number of tickets using both a binomial distribution and an
equivalent normal distribution.
}
\details{
Only checks the range [N, N*1.1]
}
\examples{
ntickets(200, 0.02, 0.95)

}
