% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zValues.R
\name{zValues}
\alias{zValues}
\title{Calculate z values of a vector}
\usage{
zValues(vector)
}
\arguments{
\item{vector}{A data vector}
}
\value{
A vector containing the Z values of the given data vector
}
\description{
Calculate z values of a vector
}
\examples{
z <- zValues(c(1, 2, 3))

}
